♻ 불필요한 Service 인터페이스 삭제
- Respository 레이어에서 이미 인터페이스를 구현하고 있으므로 Service 레이어에서는 불필요하게 다형성을 나타낼 필요가 없다.
