+++ src/test/java/com/kakao/cafe/controller/ArticleControllerTest.java
@@ -33,6 +33,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 
+@Deprecated
 @SpringBootTest
 @AutoConfigureMockMvc
 public class ArticleControllerTest {
