+++ src/test/java/com/kakao/cafe/controller/UserControllerTest.java
@@ -31,6 +31,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
+@Deprecated
 @SpringBootTest
 @AutoConfigureMockMvc
 class UserControllerTest {
